backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
easyocr
edge-tts
pydantic
numpy
Pillow
deep-translator
langdetect
gunicorn
pyngrok
aiofiles

sumy
nltk
google-generativeai
